deploy/requirements.txt
fastapi
uvicorn
databricks-sdk
httpx
